Rendering with shaders, part 6

+++ source/EngineGpuKernels/RenderingKernelsService.cu
@@ -49,10 +49,10 @@ void _RenderingKernelsService::extractObjectData(SettingsForSimulation const& se
     auto const& gpuSettings = settings.cudaSettings;
     
     // Resize buffer if necessary
-    int maxObjects = data.objects.cells.getNumEntries_host() + data.objects.particles.getNumEntries_host();
+    uint64_t maxObjects = data.objects.cells.getNumEntries_host() + data.objects.particles.getNumEntries_host();
     renderingData.resizeObjectBufferIfNecessary(maxObjects);
 
     // Extract object data
-    CHECK_FOR_CUDA_ERROR(cudaMemset(renderingData.numObjects, 0, sizeof(int)));
+    CHECK_FOR_CUDA_ERROR(cudaMemset(renderingData.numObjects, 0, sizeof(uint64_t)));
     KERNEL_CALL(cudaExtractObjectData, data.worldSize, data.objects.cells, data.objects.particles, renderingData.objectData, renderingData.numObjects);
 }
